Just a progress log
I had stopped working on this and was happy for the most part, semester
at university had started so I got busy with that work. Until for some
reason I tried making a small C hello-world just for fun. And... nothing
works? It inserts non printable characters sometimes, and saves random
things. Seems to be riddled with bugs. I don't have a lot of time at the
presnet, but I hope to clean it up one day, it's a project I enjoyed
working on and I don't want to leave it in a broken state.